symbols: bug 520635 Parse symbol file with sizes
Symbols files generated by tools like the perf-map-agent for java
generates file with a format similar to nm, but the value in the middle
represents the size of the symbol, so the default regex does not work
for those files and the size allows to delimit the region of a symbol,
symbols may contain others.

This type of mapping file uses a segment store where the start and end
are address spaces.

Change-Id: I1cdd4489d1257050eec216fb9405f2ae85da5428
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/102987
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Hudson CI