swtbot: Fix ProjectExplorerRefreshTest last modified times
Depending on the OS and file system, the last modified time can have a
granularity of 1 second (value in milliseconds always ending in 000).

In the test, if the two copyDirectory operations were executed within
the same second, the last modified time could be unchanged and the
platform would not detect a resource change on refresh.

Using a different value of the preserveFileDate parameter ensures that
the last modified time is changed and that resource change is detected.

Change-Id: Id55dee95de9e1b772613fcede0b0d6eda74656b2
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/103961
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/105105
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>