swtbot: Fix ControlFlowViewTest and ResourcesViewTest
Pressing the HOME or DOWN shortcut can return before the shortcut has
been received by the application. The test continues too early and
attempts to navigate before the proper entry has been selected.

Use the SWTBotUtils method that waits for the key events to be received.

When opening the legend, sometimes the shell is visible but it is not
active (there is no active shell or focus control at all).

Get the legend shell and activate it before continuing the test.

Change-Id: Iec3b59553a8cb4ddbe1327665639258372ee2935
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/106111
Reviewed-by: Hudson CI
Reviewed-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Tested-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>