linux.analysis: fix hanging KernelMemoryAnalysisModuleTest
testAnalysisExecution calls executeAnalysis() directly and not schedule.
This has the side effect that the dependency level of this
KernelMemoryAnalysisModule is not incremented and therefore the event
request of this module can be coalesced with TID analysis event request.
Since the kernel memory analysis uses the TID analysis
to resolve TID using an the KernelTidAspect.resolve() method with the
block=true flag, the resolve method can block forever. 

The event requests of the KernelMemoryAnalysisModule and
TidAnalysisModule must never be coalesced. Either use the schedule()
method or manually run the TidAnalysisModule and wait till it's
finished.

Change-Id: Idd3e514dd6e6381bbbbabf79cfdcb1c2245aa740
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/106256
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>