tmf.ui: Let the abstract AbstractSelectTreeViewer save checked entries.
Previously, the CounterTreeViewer handled persisting its checked entries
by persisting the tree for each trace.
This patch lets the AbstractSelectTreeViewer handle persisting the
checked entries by saving the checked entries' ids in their trace's
context, and checking the entries again when the content is updated.
This allows all TreeViewers extending AbstractSelectTreeViewer to remember
checked entries and the checked entries to be remembered even if a new
tree is returned when the time range changes.

Change-Id: I36e8db311a5a11f4428a2935051a4752abbfbe2c
Signed-off-by: Loïc Prieur-Drevon <loic.prieurdrevon@gmail.com>
Reviewed-on: https://git.eclipse.org/r/110166
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>