XML: bug 527746 Share the attribute pools across the analysis
The attribute pools were fields of an attribute, so many actions sharing
a path that contained a common attribute pool would each get a new pool,
under the same quark and would cause errors. Now this puts the attribute
pool at the state provider level, so that there is only one pool under a
given quark.

Change-Id: I820d2c753669de701d4abd39fade9ab42611f96d
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/110168
Reviewed-by: Hudson CI