symbols: Use Long.compareUnsigned to compare symbols
Symbol addresses are positive long values, they are parsed using the
Long.parseUnsignedLong method and they should be compared with
Long.compareUnsigned to avoid the large longs to be considered negative.

This patch fix the returned symbol for addresses after the last symbol
in a mapping file. Since all addresses after that last symbol would map
to that symbol, it is often wrong.

Change-Id: Ie41f5faf772b3f505d70c9999e6f19b7af00492f
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/110679
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>