os: Fix the thread on CPU on sched_switch
At the time of sched_switches, the thread on the CPU was the one
entering CPU, as the current thread changes at that time.

But at the moment of the sched_switch, the running thread should still
be the preceding one, as any counter, context, etc. apply to that one
and not the next. We should query the state system at time-1 to get the
right value at sched_switch time. It will have no impact on the returned
value for other times.

Change-Id: Ie6ddd41e529c738ca68d7324da5e9a51a69811d1
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/110682
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>