RFC: Rewrite of the XML plugin code
The XML code base, as it was, is hard to maintain, it could be simpler
as there's a lot of repetition, especially between attributes and
values.

This patch does the following:

* Create a new module to replace the XmlStateProviderModule. The helper
creates this module only if the XML validates correctly.

* Building the state provider is divided in 2 phases:

 - A compilation and semantics validation phase, which takes the XML
element, validates that elements it refers to exist and creates a data
structure for the provider made of objects called "compilation unit"
(suffix Cu to classes). It replaces elements like locations and defined
values by the appropriate value.

 - A state provider generator phase, which builds for each compilation
unit, the matching data-driven class. But when the state provider is
built, we have a guarantee that it is sound and coherent.

Change-Id: I2dbe2c0e9389f6dc2621b9be1b03aa4978b2965d
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/111398
Tested-by: Hudson CI
Reviewed-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>