tmf.ui: add alignement to TimeGraphScale
This adds the possibility of top or bottom aligning the time scale.

Note: the swt.bottom indicates that the scale is on the bottom
of the view, not that the line is on the bottom.

bug: 527430

Change-Id: I373c99382065e82806acd57b7bab300a9d668150
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/111678
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Hudson CI
(cherry picked from commit 3a24f80b292d38648a5aec2bb8b6e3e5ffd6e51e)
Reviewed-on: https://git.eclipse.org/r/112132