linux: Track process ID in kernel analysis
The ID of the process the thread is from was not tracked in the state system.
This value can be useful to resolve symbols for a thread. If the value is
absent from a thread, this thread is a process.

It also fixes the parent ID in the statedump, as the parent should be the
ppid value no matter if it is a thread or process. It now matches the behavior
of the sched_process_fork event that was assigning the parent TID as PPID.

Change-Id: Iae6327cde9ecc51594f2de98a3eb8ba03f84b3f3
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/111809
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Hudson CI