tmf.ui: make time chart views use the TimeGraphScale
This makes for a more consistent look and feel between the timegraphs
and the xy charts.

Moreover, it solves an alignment bug brought on with SWTChart v0.10.0

This patch aims to not aiming to bring the feature parity of the
timegraph view to the xy view. It aims to maintain the functionality of
the XY view. Doubleclick on the axis will reset the time range. Dragging
the timerange will do nothing as it did before.

Bug: 527430

Change-Id: I4d81f2ab327d900cc05fafe1c38bd5cbe38ec537
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/111679
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>