lttng: Bug 527674: Fix intermittend failure in DiskIOViewTest
The setting of all X-series and y-series are not atomic operations.
So, checking in the test at random times that all x-Series are the same
will fail from time to time. 

This patch won't fail the when the X-Series are not the same, but will
time out if it won't be the same within the timeout time.

Added debug message, to know what case happened when there is a timeout.

Change-Id: I81fd89dfaf7bab0c15e882cde76be472942b8c27
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/112164
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>