tmf: make the State System explorer a time graph view
The entry tree follows the structure of the state system's
attribute tree.
The time intervals' colors are representative of the values,
with transparent events for null state values for example.
The times are labeled by their value.
The analysis entries are double clickable to start the analysis
and be able to visualize its states.
The expand/collapse button is preserved.
Time synching is still supported.
Start, End, Duration, Full path, type columns are removed from
the tree as they are supplied by the tool tip in the TimeGraph
area where they are more relevant.
Updated ImportAndReadKernelSmokeTest.

Change-Id: I9ab4e7cd26edf3365e006e942bf6c6dbb9ec93ec
Signed-off-by: Loïc Prieur-Drevon <loic.prieurdrevon@gmail.com>
Reviewed-on: https://git.eclipse.org/r/112232
Reviewed-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>