tmf: Bug 527827: AbstractSelectTreeViewer checked elements lost
Override traceOpened() and traceSelected() to ensure it is executed
before the base class handles the signal.

Do not call saveViewContext() from updateTree() as that can lead to
storing the checked elements of the previous trace in the trace context
of the new trace.

Change-Id: I69117f1668359727124031cefd079bd97bbdea5d
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/112384
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
(cherry picked from commit 6282105ebd1ab3987f191046e13faf72e5d19782)
Reviewed-on: https://git.eclipse.org/r/112517