tmf: Reduce memory requirement of trace synchronization
It used the save the event, which could be up to 8K/event for
CtfTmfEvent, but only the trace and timestamp are necessary
for event matching. By keeping only this data, we reduce the
memory requirement by a lot.

The memory usage for the django benchmark has gone from 314M to 16.3M.
Incidentally, there's also a speedup of the CPU benchmark from 8.26sec
to 5.2 sec.

Change-Id: I5c917a76207192fabfbf3ba795b70c484a884f7f
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/113367
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>