ss: Fix a NPE on 2D queries
When the tree is being built, core nodes do not know the final
narrowed bounds of their children, which results in nodes that
potentially do not overlap the queried conditions being queued.
The subconditions on those nodes would then be null, as created
from the node's actual values.

This patch skips over these irrelevant nodes to avoid NPEs

Change-Id: I234b6783bfae9494460414938eadf8c845f7853a
Signed-off-by: Loic Prieur-Drevon <loic.prieurdrevon@gmail.com>
Reviewed-on: https://git.eclipse.org/r/113369
Reviewed-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>