tmf: Perform cleanup in TmfEventMatching
When all traces involved in event matching have matched events, the
older events for each trace can be cleaned up to save some memory

For the django benchmark, this reduces memory usage from 16.3M to 11.25
and also impacts positively the time of event matching.

Drawback: this cleanup approach will not work if some trace hosts in the
experiment do not have matching events. Also it does not discriminate by
matching type, so if there are both TCP packets (low latency) and some UST
event matching (with potentially much higher latencies), the cause of the
UST event may have been cleaned up when comes the effect. It is ideal for
trace synchronization but other uses of event matching may not like it.

Change-Id: I95666c06e13d6868675c5b9ade25996dd83a6d78
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/113435
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>