timing: Add an event matching latency analysis
This analysis shows the latency between 2 events using the event
matching mechanism. It can be used to visually see how accurate event
matching is through time, before or after synchronization.

The EventMatchingLatency is a INamedSegment and the name is the
type of matching used, followed by the 2 traces involved in the
match

Change-Id: Idbe31e438ab2f0807112dd4c44d7f3d50b6b6a3d
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/113450
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Hudson CI