tmf: Bug 528919: Only call System.gc() when needed
When a trace is closed we set a marker flag so that the GC is called
just before deleting the trace. This avoids calling GC for every
trace that is being deleted.

A flag was introduced keep track if a trace has been previously
closed. This flag is checked before deleting a trace and invoke
the System.gc() to free MappedByteBuffers. This is to workaroun
Java bug JDK-4724038.

Change-Id: I9845544f403c4c77c1c39476f4eea2e0b6d4aac7
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/113672
Reviewed-by: Loic Prieur-Drevon <loic.prieurdrevon@gmail.com>
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>