lttng2.swtbot: fix MemoryUsageViewTest intermittent failure
The test verified that the view was ready by selecting all tree
entries and ensuring that the expected number of series were loaded.
However when the selection could happen before the full tree was
ready, therefore an incomplete set of series would be queried.

This patch ensures that the complete tree is loaded before selecting
and querying series.

Change-Id: If90eb76dff6f72432788620cbff8540f6a323609
Signed-off-by: Loic Prieur-Drevon <loic.prieurdrevon@gmail.com>
Reviewed-on: https://git.eclipse.org/r/116112
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Hudson CI