tmf: bug 531417 Fix infinite loop in composite provider when some fail
When the window range for an experiment is a time range where on the traces
does not have data, then the data provider returns a failed response for
this trace. And since the view considers the work complete only when all
traces have a completed response, then it loops forever in the fetch data.

This fixes this by considering a query incomplete only if at least one
provider is still running.

Change-Id: I1bdc218bba0cd4772108d780c90b97476bfc5322
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/116122
Tested-by: Hudson CI
Reviewed-by: Loic Prieur-Drevon <loic.prieurdrevon@gmail.com>