core: Add a method to list available providers for a trace.
Adds a method to IDataProviderFactory to return a list of descriptors
for the trace, if given trace supports the provider.
The provider contains the an unique ID, a name, the type of provider
(TABLE, TREE_TIME_XY, or TIME_GRAPH) and a short description of
the data returned by the provider.

Not supported in this patch:
- Events table
- Segment store statistics tree
- XML time graph, XML XY and tree

Change-Id: Ia334896b98036057ac6007a700a268aa557250a9
Signed-off-by: Loic Prieur-Drevon <loic.prieurdrevon@gmail.com>
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/135899
Tested-by: CI Bot
Reviewed-by: Simon Delisle <simon.delisle@ericsson.com>
Tested-by: Simon Delisle <simon.delisle@ericsson.com>