tmf.core: add a thread-safe method to get queried entries in providers
Previously, the AbstractTreeDataProvider only provided a method to
get selected quarks from a TimeQueryFilter (thread-safe)
and another method to get selected Ids from their respective quarks
(not thread-safe).

This patch adds a thread-safe method to retrieve a BiMap of
selected entries' Ids to their respective quarks, avoiding having to
go back and forth with the other methods and adding safety.

Change-Id: Ia3634d84e334609488fc8c9d66f530ecbd122a80
Signed-off-by: Loïc Prieur-Drevon <loic.prieurdrevon@gmail.com>
Reviewed-on: https://git.eclipse.org/r/116652
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Hudson CI