tmf: Improve speed of Clear, Delete and Remove trace folder elements
The list of traces and experiments to close is compiled in the
background. To prevent possible deadlock, the editors are closed in a
loop outside of any workspace modify operation.

The list of resources to delete is compiled in the background, sorted by
project. Duplicates are removed so that a child resource does not need
to be explicitly deleted if any of its parent resources is already
selected for deletion.

A separate workspace modify operation is then run for each affected
project, allowing a more fine grained lock on the workspace.

When deleting traces and trace folders, the reduced set of resources are
first deleted from the workspace, and traces are then deleted, which
cleans their supplementary files and remove the traces from experiments
to which they belong. The trace resources are then deleted but this is a
no-op since it has already been deleted.

The deletion of experiments is no longer done on the UI thread.

TmfTraceElement.delete() is fixed to prevent deletion of supplementary
files from TmfProjectRegistry.handleTraceDeleted() when overwriting a
trace (e.g. during trace package import).

TmfProjectModelElement.refreshViewer() is fixed when called on a shadow
project element.

TmfProjectRegistry.resourceChanged() is fixed to properly handle deleted
trace resources that are files (e.g. not CTF folders).

TmfProjectRegistry.resourceChanged() is fixed to properly clean up
supplementary files when an experiment trace is deleted directly from
the file system.

TmfProjectRegistry.findElement() is fixed to be able to find experiment
traces.

Change-Id: Iab174be610d4324a45108bf24cd3f14e0eb695d5
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/116998
Tested-by: CI Bot
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>