tmf: Fix experiment deep copy
When creating a deep copy of an experiment, the folder structure under
the copied experiment was moved to a new subFolder with the name of the
new copied experiment. This was done in a loop for every experiment
trace, so if a second experiment trace was in the same original root
folder, it no longer existed when it was attempted to be moved again.

This caused an exception that was silently caught, and prevented the
remaining traces from having their link properly updated or their folder
structure to be moved.

The new solution deletes all original traces from the copied experiment
then adds the copied traces. This needs to be done without having the
copied trace and/or experiment elements created to avoid having
supplementary files deleted.

Change-Id: I6bf02078726903d4c5b3f66a7e48497a82fae2f0
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/117305
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>