tmf: fix intermittent call stack view tests
When selecting a time in the events table, it's not sufficient to 
wait until the events table selects the right time. The test needs also
to wait that the Call Stack view selects the correct time before
continuing with the test case. The SWTBot test
CallStackViewTest.testGoToTimeGoBackAndForthAndCheckStack() uses the
next/previous button and then verifies that the events table is
updated. Before using the button the first time the test uses the
events table to navigate to a certain time. But it doesn't wait for
the Call Stack view to seek to that time.

Change-Id: Ib4cc7d19bc22a35cd4e3df905ee8de9c75f8a043
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/117560
Tested-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>