tmf: Bug 531357: TmfProjectRegistry opens BlockedJobsDialog
If an atomic workspace operation is run from the UI thread with
IWorkspace.run(), then the BlockedJobsDialog can be shown while the
workspace is locked by another workspace operation.

When running a background task, IWorkspace.run() should not be called
from the UI thread.

The TmfProjectRegistry resource change listener will instead run the
atomic workspace operation from a system job.

For deleted traces, a single syncExec runnable will close all necessary
editors first. The workspace operation will then delete all required
resources without going to the UI thread.

For changed traces, execution will be deferred to an asyncExec runnable
only when necessary to prompt the user and/or close editors, and if
required, a separate system job will be scheduled to delete the
supplementary resources in an atomic workspace operation.

Change-Id: I5ea6d45ee9bccffc2851e84cadcc3bf8f0779196
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/117692
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Hudson CI