tmf: Window range update from selection update may be missed
In AbstractTimeGraphView, when the TmfSelectionRangeUpdatedSignal is
received, the view should ensure that the selected time is visible.

However the signal may be received from an other trace that should not
synchronize with the current trace. TmfTraceManager takes care of this
by ignoring the selection range update. The view then aborts the
selection update (and ensure visible) if the current selection is the
same as the selection the TmfTraceManager context.

This leads to two cases where a required window range update (to ensure
selection is visible) may be missed:

- If the user navigates away from the selection, then re-selects the
same time (e.g. in the event table), the window range doesn't come back
to the selection.

- If a spontaneous refresh occurs in AbstractTimeGraphView while the
signal is be handled, the refresh can synchronize the selection with the
TmfTraceManager context (without ensure visible). When the asyncExec
runnable from the signal handling is executed, the view aborts the
update because the selection is already equal to the updated selection.

To fix this, an API is added in TmfTraceManager to indicate if a trace
should be synchronized with an other trace. If it should, then the
selection update (and ensure visible) is executed regardless of whether
the selection has changed or not. Otherwise the handling of the signal
is aborted early.

Change-Id: If67e8d3f6b627ea87bcbaa5b3eaf2f4ccaf47adb
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/117823
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>