tmf: Bug 531558: Opening trace can freeze UI during workspace operation
Use IProgressService.busyCursorWhile() to refresh the supplementary
folder and create the bookmarks file from the UI thread when opening a
trace.

Pass the provided progress monitor to the refresh supplementary folder
and bookmark creating methods to allow the progress dialog to be shown
if the operation is blocked for a long time.

Cancel opening the trace if the operation is canceled by the user.

Change-Id: Ia1ae89564eea4806c7259ab513b39ce2d739cb48
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/117984
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>