tmf: Make experiment traces persisted as empty files
Instead of linked resources, experiment traces are now persisted as
empty files. The file's relative path to the experiment corresponds to
the trace element's relative path to the Traces folder.

This greatly improves performance for projects with a large number of
experiment traces, as linked resources are persisted in the platform in
the .project file which is in XML format and can grow very large,
requiring increasingly long time to read and write when adding, removing
or modifying linked resources.

Experiment traces as linked resources are still supported for backward
compatibility with older workspaces but the change is not forward
compatible. Experiments created in the new way will appear to have no
traces if the workspace is used with an older version of the framework.

Experiment traces now have the link decorator only if the corresponding
trace under the Traces folder is itself a link.

The method getElementUnderTraceFolder() is made more efficient by
walking down the element path instead of getting all traces in the
Traces folder recursively and searching for the correct one.

Change-Id: Id23a3db92fc0751a186e4a57ef0d7bd35c7a99b3
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/118088
Tested-by: CI Bot
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>