core: Support data providers with optional secondary IDs
This patch adds a method to retrieve providers by ID and secondary ID
(typically the analysisID, but factories decide what this ID may mean).
A new method in factories allows to create the provider by giving it a
secondary ID.

This will allow the same data provider to be used for many analyses,
while not having to do a separate way of retrieving the providers.

Change-Id: Ie995479df0676e76b5586850da70bd00ce9d1ad2
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/119336
Tested-by: CI Bot
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>