tmf: Fix window range of empty traces in AbstractTimeGraphView
The intention of the changed code was to set the window range to
[SWT.DEFAULT, SWT.DEFAULT] when the time bounds were never set, as
indicated by the invalid bounds [Long.MAX_VALUE, Long.MIN_VALUE].

However when the trace was empty, its time range and consequently the
view's time bounds could be set to [Long.MIN_VALUE, Long.MIN_VALUE]. The
window range would then interpret this as if only the upper bound was
never set, and set the window range to [Long.MIN_VALUE, SWT.DEFAULT].

The code now checks both bounds to determine if the time bounds were
never set. It will also use the default window range for any invalid
time bounds (start > end).

Also, in the case that they were set and valid, the trace manager's
window range is now clamped to both the upper bound and the lower bound
of the time bounds, for both the start and the end of the window range.

Change-Id: Id1d66f9eff729999e8a903172069b2d1571c177d
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/119903
Tested-by: Loic Prieur-Drevon <loic.prieurdrevon@gmail.com>
Tested-by: CI Bot
Reviewed-by: Loic Prieur-Drevon <loic.prieurdrevon@gmail.com>