swtbot: Explicitly activate new shells
For some reason, sometimes a new shell that is opened by the application
does not get activated. Instead of waiting for the new shell to become
active, tests should get the shell by title and explicitly activate it.

Tests should also avoid using activeShell() if they can get a reference
to the shell some other way.

Change-Id: I8d92db1987e8ad0a4507ec62325e2f0c2c00a476
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/119921
Tested-by: CI Bot
Reviewed-by: Loic Prieur-Drevon <loic.prieurdrevon@gmail.com>