swtbot: Fix XmlTimegraphViewTest
WaitUtils.waitForJobs() ensures that the build thread is complete, but
does not ensure that the triggered refresh on the UI thread has been
executed, as it is queued in the TmfUiRefreshHandler.

We can use SWTBotTimeGraphEntry.getEntry(String) as this will wait until
the desired entry is found.

The window range is also checked using SWTBotUntils.waitUntil(), in case
there is a delay.

Change-Id: I72c5222bbe2679e9bdd0a7ea38d720a8d3018ea2
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/120193
Tested-by: CI Bot
Reviewed-by: Loic Prieur-Drevon <loic.prieurdrevon@gmail.com>
Tested-by: Loic Prieur-Drevon <loic.prieurdrevon@gmail.com>