tmf: bug 532971 Make smarter cast in ITmfEventField#getFieldValue
When the field is not directly assignable from the requested type, the
method can try to cast between some known types. This allows to
transparently cast a field that comes as a String to a numeric type.

This patch also changes some usages of the previous getField method to
getFieldValue in the linux analysis.

Change-Id: Idecd8d0db79a1c3f0bc9c73d5afa9423fbe8c161
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/120365
Tested-by: CI Bot
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>