swtbot: Fix ExportToTsvTest
The event table's context menu is dependent on the table selection. The
test should make sure that a table item other than the header row is
selected when invoking the context menu, by invoking it on a specific
table item.

Change confusing test names and descriptions that were copied from
CopyToClipboardTest. Unlike Copy To Clipboard, the Export To Text result
is affected by whether a filter is applied or not, it does not depend on
the current table selection.

Change-Id: I33e427a9c5b7d5c2571808d40465756a51288818
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/120471
Tested-by: Loic Prieur-Drevon <loic.prieurdrevon@gmail.com>
Tested-by: CI Bot
Reviewed-by: Loic Prieur-Drevon <loic.prieurdrevon@gmail.com>