tmf.ui: Use data provider for histogram
The provider is now used to populate the histogram, currently it
populates the histogram data model with the same data as the event
request before. That means that the model still need to be scaled even
if the provider support returning the right data.
Future improvement: investigate with level we can remove to avoid
scaling the model and use the provider more directly.

Change-Id: I8d6d195c5094df5aa694d57642f2d6cec1578b4c
Signed-off-by: Simon Delisle <simon.delisle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/120652
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: CI Bot