tmf: Add AutoFit and Reset All actions to event table column header menu
When a column has AutoFit enabled, it will automatically increase its
width to fit the cell contents. The width will never decrease.

AutoFit is disabled when a column is resized manually by the user.

AutoFit can be toggled on or off for a specific column by selecting the
'AutoFit' check action in its table column header context menu.

AutoFit is initially enabled by default for all columns, and its state
is persisted (as a column width set to SWT.DEFAULT for enabled state)
when the trace is closed.

The 'Reset All' action is added to the table column header context menu.
It resets the column order to default, resets the column visibility
according to ITmfEventAspect.isHiddenByDefault(), and resets AutoFit to
enabled for all columns.

Method TmfVirtualTable.getColumn(Point) is added to help determine the
column on which the table header context menu was detected.

Methods getHorizontalBar() and getVerticalBar() are overridden in
TmfVirtualTable to return the Table's scroll bars instead of the virtual
table Composite's.

Change-Id: I0eb0bb6cc443e1232fbf1ddb48c8d1c7e0503978
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/120870
Tested-by: CI Bot
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>