os: bug 532973 Add a CPU frequency line
This patch adds weighted model entry type that gives hint to the UI to
show the values as a percentage between a minimum and maximum value.

It adds the CPU frequency event to the layout and adds the CPU
frequency, if available, for each CPU.

The view displays the CPU frequency as a weighted value between a given
CPU's minimum and maximum value.

Change-Id: Ibce6dd344eb74011be835e305ca54ee02114b805
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/121806
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: CI Bot