tmf: Bug 532127: Improve drawing of time graph states
The drawing of time graph states is changed so that the next state never
overlaps with the previous state's full width. The state start will be
truncated, and if it is completely overlapped by the previous state,
only a dot will be drawn above the state's start pixel.

The drawing of transparent states will now fill the background with a
transparent light gray color. This will ensure that short transparent
states between null states will be visible in the time graph instead of
only showing a dot above blank space. Transparent states will never
overlap to ensure that the transparent color is not blended with itself.

Change-Id: I0586fff9f2f8252ecbb471d88bf5560cbb2e6453
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/121972
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: CI Bot