Change data providers API from provisional to public API
Move data provider classes from internal provisional to internal and
public. All the interfaces and models used to create data provider are
now API and all the implementations of these interfaces will remain
internal for the moment. When we feel comfortable and providers are more
widely used, we can move the "abstract" implementation to public.

Change-Id: I69d74e9f8579945e311f799a1697cd1ef301c282
Signed-off-by: Simon Delisle <simon.delisle@ericsson.com>
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/122045
Tested-by: CI Bot
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>