xml.ui: add enabling and disabling of XML analyses files
Also, add a SWTBot method for opening the preferences dialog by
providing an alternative text for the dialog, in case
the preference page is not opened for the first time.

Also, add a SWTBot method for getting the number of checked items
in a table, similar to the tree version.

Change-Id: Ie6f8d88ffa86d35bb56dc2501161b8b821d3a553
Signed-off-by: Christophe Bourque Bedard <christophe.bourque.bedard@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/122484
Reviewed-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Tested-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Tested-by: CI Bot
Reviewed-by: Simon Delisle <simon.delisle@ericsson.com>