tmf: Bug 534393: BaseDataProviderTimeGraphView broken by non-unique ids
- Change BaseDataProviderTimeGraphView to use a table of (data provider,
model id) to time graph entries.

- Reuse of same model id for different time graph entries is no longer
allowed. The mapping of model id to time graph entry is no longer a
multimap.

- ResourcesStatusDataProvider now keeps a BiMap of twin entry model id
to corresponding quark, used when fetching the row models.

- AbstractTreeDataProvider.getSelectedEntries() now returns a Map
instead of a BiMap, so that the returned map can be augmented with
mapping for a model id to a quark that is already present as a value for
another model id in the map.

- BaseDataProviderTimeGraphView clears the table when a trace is closed.

- BaseDataProviderTimeGraphView.setLinkList() now correctly uses the
arrow value from the model when creating a TimeLinkEvent.

- ControlFlowView now fills and uses the base class's time graph entry
table instead of keeping its own table. The method getLinkList() is no
longer overridden.

- XmlTimeGraphView now fills and uses the base class's time graph entry
table.

Change-Id: I865032eef04aa080baa006d4055ee26d8c5a867f
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/122210
Tested-by: CI Bot
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>