tmf: Bug 534773: Sampling gaps not visible in XmlTimeGraphView
Sampling gaps (TimeEvent with no value) are now drawn using the
TRANSPARENT state instead of being mapped to the INVISIBLE state.

Unknown values in state system are now mapped to TimeGraphState with
value Integer.MIN_VALUE, which is then mapped to a NullTimeEvent and to
INVISIBLE state.

The duration of each TimeGraphState is fixed to take into account the -1
offset in ITmfStateInterval.getEndTime().

Change-Id: I94614f8440b52fac95024846c70c060ed8c00361
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/122806
Tested-by: CI Bot
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>